pyproject.toml
[tool.poetry]
name = "twitter-analytics"
version = "0.1.0"
description = "Aplikasi ini diperuntukan untuk pembelajaran"
authors = ["Oktapian Candra <2211510678@student.budiluhur.ac.id>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.9.0"
tweepy = "^4.14.0"
module-typica = {version='0.1.4', source='ebdesk'}
pydantic = "^2.5.1"
pydantic-settings = "^2.1.0"
loguru = "^0.7.2"
requests = "^2.31.0"
redis = "^5.0.1"
pandas = "^2.1.4"
scikit-learn = "^1.3.2"

[tool.poetry.scripts]
colleges = "src.main:app"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[[tool.poetry.source]]
name = "ebdesk"
url = "http://192.168.20.127:6060/simple"
priority = "supplemental"
